third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-origin-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-origin with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-origin">
<meta name="assert" content="mask-origin supports the full grammar '<geometry-box>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>

// <geomatry-box>
test_valid_value("mask-origin", "border-box");
test_valid_value("mask-origin", "padding-box");
test_valid_value("mask-origin", "content-box");
test_valid_value("mask-origin", "margin-box");
test_valid_value("mask-origin", "fill-box");
test_valid_value("mask-origin", "stroke-box");
test_valid_value("mask-origin", "view-box");

test_valid_value("mask-origin", "border-box, content-box");
</script>
</body>
</html>
